requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==20.2.0
chardet==3.0.4
idna==2.10
ldap3==2.8.1
multidict==4.7.6
pyasn1==0.4.8
PyJWT==1.7.1
uvloop==0.14.0
yarl==1.6.0
